fixed bugs of position and add intron variants extraction
1.   For the missing annotation from annovar output, (see the commit
comment in the previous commit), I found that the possible reason is the
unknown bug of multithreading in annovar. I ran again the same script
switching off the multithreading argument. Foturnatively, it worked
again ! Oh, well, but the modification from the past  are still kept
here, in case of the next hash~~

2.   In prepare the GRange object,, we have to set the argument
starts.in.df.are.0based = FALSE. Because before this we have transformed
the coordination from 0-based to the 1-based one.

3.   For comparison of different setting on running MutsibCV, I prepared
two set of variants, one using only exonic+UTR variants, the other
iinvolving intron variants. I expected the latter one is more close to
what the author claimed.